rename createExplorer*
