+++ src/ExplorerBase.ts
